git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@5513 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Include/WorkingBlockHeader.h
@@ -29,7 +29,7 @@ typedef struct {
 #define WORKING_BLOCK_INVALID 0x2
   UINT8     Reserved : 6;
   UINT8     Reserved3[3];
-  UINT32    WriteQueueSize;
+  UINT64    WriteQueueSize;
   //
   // UINT8                WriteQueue[WriteQueueSize];
   //
